+++ app/src/main/java/com/example/pocastcloni/util/Constants.kt
@@ -101,7 +101,7 @@ object Constants {
 
     object Database {
         const val DATABASE_NAME = "pocast_cloni_db"
-        const val DATABASE_VERSION = 17
+        const val DATABASE_VERSION = 18
         const val DEFAULT_SORT_ORDER = 0L
         const val TABLE_PODCASTS = "podcasts"
         const val TABLE_EPISODES = "episodes"